chore: put friend functionality in its own files for clarity and clean folder structure
