+++ backend/user_management/src/userRoutes.ts
@@ -9,18 +9,11 @@ import {
 	insertUserHandler,
 	setStatusHandler,
 	getUserIdByUsernameHandler,
-	getUsernameByUserIdHandler,
-
-	friendRequestHandler,
-	acceptFriendRequestHandler,
-	rejectFriendRequestHandler,
-	getFriendRequestsHandler,
-	getFriendsHandler,
-	removeFriendHandler
+	getUsernameByUserIdHandler
 } from "./userController.js";
 
 // prefix: /users
-export const userRoutes = async (app: FastifyInstance) => {
+const userRoutes = async (app: FastifyInstance) => {
 	app.get('/profile',		{ preHandler: [app.authenticate] },	getUserDataHandler);
 	app.put('/username',	{ preHandler: [app.authenticate] },	updateUsernameHandler);
 	app.put('/password',	{ preHandler: [app.authenticate] },	updatePasswordHandler);
@@ -33,14 +26,4 @@ export const userRoutes = async (app: FastifyInstance) => {
 	app.get('/get-username',	getUsernameByUserIdHandler);
 };
 
-// prefix: /users/friends
-export const friendRoutes = async (app: FastifyInstance) => {
-	app.post('/request',	{ preHandler: [app.authenticate] }, friendRequestHandler);
-	app.put('/accept',		{ preHandler: [app.authenticate] }, acceptFriendRequestHandler);
-	app.put('/reject',		{ preHandler: [app.authenticate] }, rejectFriendRequestHandler);
-	
-	app.get('/requests',	{ preHandler: [app.authenticate] }, getFriendRequestsHandler);
-	app.get('/list',		{ preHandler: [app.authenticate] }, getFriendsHandler);
-	
-	app.delete('/remove',	{ preHandler: [app.authenticate] }, removeFriendHandler);
-};
+export default userRoutes;
